refactor : use refactored code
