+++ src/components/container/WritePageContainer.tsx
@@ -4,7 +4,7 @@ import CategoryInput from 'src/components/Inputs/CategoryInput';
 import TitleInput from 'src/components/Inputs/TitleInput';
 import styled from '@emotion/styled';
 
-import { useArticleElement } from 'src/hooks/useHandleArticle';
+import { useArticleElement } from 'src/hooks';
 
 const WritePageContainer = () => {
   return (
