+++ src/components/modals/create-article-Modal/CreateArticleModal.tsx
@@ -9,15 +9,15 @@ import DescInput from 'src/components/Inputs/DescInput/DescInput';
 import TagInput from 'src/components/Inputs/TagInput/TagInput';
 
 import {
-  useResolveSaveFunction,
+  useHandleOpenModal,
+  useResolveSaveHandler,
   useArticleElement,
-} from 'src/hooks/useHandleArticle';
-import { useHandleOpenModal } from 'src/hooks/useHandleOpenModal';
+} from 'src/hooks';
 import { articleValidationSecondStep } from 'src/recoil/article';
 import { useRecoilValue } from 'recoil';
 
 const CreateArticleModal = () => {
-  const handleSaveArticle = useResolveSaveFunction();
+  const handleSaveArticle = useResolveSaveHandler();
   const handleOpenModal = useHandleOpenModal();
   const isValid = useRecoilValue(articleValidationSecondStep);
