+++ src/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from './usePreview';
-export * from './useHandleArticle';
 export * from './useDetectResolution';
 export * from './useHandleOpenModal';
 export * from './useViewportWidth';
+export * from './useArticleElement';
+export * from './article/useHandleArticle';
